requirements.txt
pandas
selenium
selenium-requests
webdriver-manager
pyspark
pytest
apache-airflow-providers-apache-spark
